requirements.txt
streamlit
langchain
langchain-aws
boto3
python-dotenv
pydantic
reportlab
pillow